CMakeLists.txt: set(CMAKE_CXX_STANDARD 98).
Currently, GCC generates a massive warning with over 60 lines
for a single class, the last few lines are:

    fpoptimizer.cc:768:13: warning: implicitly-declared
    ‘FPoptimizer_Optimize::MatchInfo<double>::MatchInfo(const
    FPoptimizer_Optimize::MatchInfo<double>&)’ is deprecated
    [-Wdeprecated-copy]
      768 | #define yD1 AnyParams_Rec
          |             ^~~~~~~~~~~~~

This is a legitimate warning about the potential removal of a
deprecated default behavior by future C++ versions.

Since fparser is already a legacy project, and the problem exists
in auto-generates code, it's unlikely to be fixed any time soon,
fixing it would be a very-low priority task for openEMS. For now,
we simply use "set(CMAKE_CXX_STANDARD 98)" in "CMakeFiles.txt" to
restrict ourselves to C++98 only, removing the warning and ensuring
its behavior won't change for any future compiler. Optionally,
fparser has C++11 features, but none is enabled by default.

Note that the flag CMAKE_CXX_STANDARD is introduced in CMake 3.1,
not CMake 3.0, we bump the minimum CMake version accordingly.
Since CMake 3.5 is available even on Ubuntu 14.04, compatibility
issues are unlikely.

Signed-off-by: Yifeng Li <tomli@tomli.me>